Use OAuth via Net::Github
This creates a Net::Github object authorized with OAuth in Github.pm
for use in the various subclasses. None of the subclasses have been
changed to use it yet.

Github appears to use OAuth to generate a single token (instead of
the consumer/access key/secret quadruplet Twitter does). The user
will be prompted for their password if they do not already have
such a token. The token is stored in ~/.dzil/github.ini.

Begins fixing #11